Revert "עדכון C#"
This reverts commit b3098e8ccc2c9f8510e74fd5c4c982b6a5e7c241.

+++ BLL/Convertion/CitiesConvertion.cs
@@ -31,7 +31,7 @@ namespace BLL.Convertion
             {
                 Id = city.Id,
                 Name = city.Name,
-                AreaId = (int)city.AreaId
+                AreaId = city.AreaId
             };
         }
         public static List<city> Convert(List<CityDTO> city)
